Added terra switching functionality.

+++ Project Summoner/Assets/Scripts/Battle System Scripts/Battle HUD Scripts/PartyBanner.cs
@@ -45,7 +45,7 @@ public class PartyBanner : MonoBehaviour
 
         Button partyBannerBtn = GetComponent<Button>();
         partyBannerBtn.onClick.AddListener(delegate {
-            optionSelectionUI.OpenOptionSelctionUI(terra, terraPartyIndex, battleSystem);
+            optionSelectionUI.OpenOptionSelctionUI(terra, (int)terraPartyIndex, battleSystem);
         });
     }
